+++ app/utils.js
@@ -1,3 +1,7 @@
+import fs from 'fs';
+import path from 'path';
+import {StringDecoder} from 'string_decoder';
+const decoder = new StringDecoder('utf8');
 import axios from 'axios';
 import _ from 'lodash';
 import state from './state';
@@ -127,6 +131,105 @@ export var toHex = (str, totalChars)=>{
   str = ('0'.repeat(totalChars)+Number(str).toString(16)).slice(-totalChars).toUpperCase();
   return str;
 }
+
+export var walk = (dir, done)=>{
+  var results = [];
+  fs.readdir(dir, (err, list)=>{
+    if (err) {
+      return done(err);
+    }
+    var pending = list.length;
+    if (!pending) {
+      return done(null, results);
+    }
+    _.each(list, (file)=>{
+      file = path.resolve(dir, file);
+      fs.stat(file, (err, stat)=>{
+        if (stat && stat.isDirectory()) {
+          walk(file, (err, res)=>{
+            results = results.concat(res);
+            if (!--pending) {
+              done(null, results);
+            }
+          });
+        } else {
+          results.push(file);
+          if (!--pending) {
+            done(null, results);
+          }
+        }
+      });
+    });
+  });
+};
+
+export var getLastGameModeSave = (saveDirectory, mode, cb)=>{
+  return new Promise((resolve, reject)=>{
+    walk(saveDirectory, (err, results)=>{
+      if (err) {
+        console.log(err)
+        reject(err);
+      }
+      results = _.filter(results, (result)=>{
+        return result.indexOf('st_') !== -1 && result.indexOf('\\cache\\') === -1 && result.indexOf('.hg') !== -1 && result.indexOf('mf_') === -1;
+      });
+
+      let obj = {
+        normal: [0, 1, 2],
+        survival: [3, 4, 5],
+        creative: [6, 7, 8],
+        permadeath: [9, 10, 11]
+      };
+      let saves = [];
+      let saveInts = obj[mode];
+      _.each(saveInts, (int)=>{
+        _.each(results, (result)=>{
+          let fileName = _.last(result.split('\\'));
+          if (int === 0 && fileName === 'storage.hg' || result.indexOf(`storage${int}.hg`) !== -1) {
+            saves.push({
+              fileName: fileName,
+              result: result,
+              mtime: fs.statSync(result).mtime
+            });
+          }
+        });
+      });
+
+      let lastModifiedSave = _.chain(saves).orderBy('mtime', 'asc').last().value();
+      if (!lastModifiedSave) {
+        reject();
+        return;
+      }
+      lastModifiedSave.path = lastModifiedSave.result;
+      lastModifiedSave.result = decoder.write(fs.readFileSync(lastModifiedSave.result)).replace(/\0$/, '');
+      lastModifiedSave.result = JSON.parse(lastModifiedSave.result);
+      resolve(lastModifiedSave);
+    });
+  });
+};
+
+export var each = (obj, cb)=>{
+  if (Array.isArray(obj)) {
+    for (let i = 0, len = obj.length; i < len; i++) {
+      cb(obj[i], i);
+    }
+  } else {
+    for (let key in obj) {
+      cb(obj[key], key);
+    }
+  }
+};
+
+export var writeCurrentSaveFile = (fileName, json, cb)=>{
+  json = JSON.stringify(json) + '\0';
+  fs.writeFile(fileName, json, {flag: 'w'}, (err, data)=>{
+    if (err) {
+      cb(err);
+      return;
+    }
+    cb(null);
+  });
+};
 
 export var css = (styleObject, newObject)=>{
   return _.assignIn(_.clone(styleObject), _.clone(newObject));